Huge refactor, and save use savedata to RL profile so it can be loaded on other PCs
