+++ src/main/java/com/logmaster/LogMasterPlugin.java
@@ -1,42 +1,26 @@
 package com.logmaster;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonObject;
-import com.google.gson.reflect.TypeToken;
 import com.google.inject.Provides;
-import com.logmaster.domain.SaveData;
 import com.logmaster.domain.Task;
 import com.logmaster.domain.TaskPointer;
 import com.logmaster.domain.TaskTier;
-import com.logmaster.domain.TieredTaskList;
-import com.logmaster.ui.UIButton;
-import com.logmaster.ui.UICheckBox;
-import com.logmaster.ui.UIComponent;
-import com.logmaster.ui.UIGraphic;
-import lombok.Getter;
-import lombok.NonNull;
+import com.logmaster.persistence.SaveDataManager;
+import com.logmaster.task.TaskService;
+import com.logmaster.ui.InterfaceManager;
+import com.logmaster.ui.component.TaskOverlay;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.ChatMessageType;
 import net.runelite.api.Client;
 import net.runelite.api.GameState;
-import net.runelite.api.IndexedSprite;
-import net.runelite.api.ItemComposition;
-import net.runelite.api.MessageNode;
-import net.runelite.api.Player;
 import net.runelite.api.SoundEffectID;
-import net.runelite.api.events.ChatMessage;
 import net.runelite.api.events.GameStateChanged;
 import net.runelite.api.events.GameTick;
 import net.runelite.api.events.WidgetClosed;
 import net.runelite.api.events.WidgetLoaded;
 import net.runelite.api.widgets.Widget;
 import net.runelite.api.widgets.WidgetInfo;
-import net.runelite.api.widgets.WidgetType;
-import net.runelite.client.RuneLite;
 import net.runelite.client.callback.ClientThread;
-import net.runelite.client.chat.ChatColorType;
 import net.runelite.client.chat.ChatCommandManager;
-import net.runelite.client.chat.ChatMessageBuilder;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.eventbus.Subscribe;
 import net.runelite.client.events.ConfigChanged;
@@ -47,66 +31,21 @@ import net.runelite.client.input.MouseWheelListener;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
 import net.runelite.client.ui.overlay.OverlayManager;
-import net.runelite.client.util.ImageUtil;
-import net.runelite.client.util.Text;
-import okhttp3.Call;
-import okhttp3.Callback;
-import okhttp3.Response;
 
 import javax.inject.Inject;
 import java.awt.event.MouseWheelEvent;
-import java.awt.image.BufferedImage;
 import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.PrintWriter;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Scanner;
-import java.util.Set;
 import java.util.stream.Collectors;
-
-import static net.runelite.http.api.RuneLiteAPI.GSON;
 
 @Slf4j
 @PluginDescriptor(
 	name = "Collection Log Master"
 )
-public class LogMasterPlugin extends Plugin implements MouseWheelListener
-{
-	public static final String DEF_FILE_SPRITES = "SpriteDef.json";
-	public static final String DEF_FILE_TASKS = "default-tasks.json";
-	public static final int TASK_BACKGROUND_SPRITE_ID = -20006;
-	public static final int TASK_LIST_BACKGROUND_SPRITE_ID = -20012;
-	public static final int TASK_COMPLETE_BACKGROUND_SPRITE_ID = -20013;
-	public static final int TASK_CURRENT_BACKGROUND_SPRITE_ID = -20016;
-
-	private static final String DATA_FOLDER_NAME = "generate-task";
-	public static final int COLLECTION_LOG_WINDOW_WIDTH = 500;
-	public static final int COLLECTION_LOG_WINDOW_HEIGHT = 314;
-	public static final int COLLECTION_LOG_CONTENT_WIDGET_ID = 40697858;
-
-	private static final int DASHBOARD_TAB_SPRITE_ID = -20007;
-	private static final int DASHBOARD_TAB_HOVER_SPRITE_ID = -20008;
-	private static final int TASKLIST_TAB_SPRITE_ID = -20009;
-	private static final int TASKLIST_TAB_HOVER_SPRITE_ID = -20010;
-	private static final int DIVIDER_SPRITE_ID = -20011;
-
-	public static final int TASKLIST_EASY_TAB_SPRITE_ID = -20017;
-	public static final int TASKLIST_EASY_TAB_HOVER_SPRITE_ID = -20018;
-	public static final int TASKLIST_MEDIUM_TAB_SPRITE_ID = -20019;
-	public static final int TASKLIST_MEDIUM_TAB_HOVER_SPRITE_ID = -20020;
-	public static final int TASKLIST_HARD_TAB_SPRITE_ID = -20021;
-	public static final int TASKLIST_HARD_TAB_HOVER_SPRITE_ID = -20022;
-	public static final int TASKLIST_ELITE_TAB_SPRITE_ID = -20023;
-	public static final int TASKLIST_ELITE_TAB_HOVER_SPRITE_ID = -20024;
-	public static final int TASKLIST_MASTER_TAB_SPRITE_ID = -20025;
-	public static final int TASKLIST_MASTER_TAB_HOVER_SPRITE_ID = -20026;
+public class LogMasterPlugin extends Plugin implements MouseWheelListener {
 	private static final String TASK_CHAT_COMMAND = "!tasker";
 
 
@@ -119,11 +58,6 @@ public class LogMasterPlugin extends Plugin implements MouseWheelListener
 	@Inject
 	private LogMasterConfig config;
 
-	@Inject
-	TaskListClient taskListClient;
-
-	@Inject
-	private Gson gson;
 
 	@Inject
 	private SpriteManager spriteManager;
@@ -141,40 +75,26 @@ public class LogMasterPlugin extends Plugin implements MouseWheelListener
 	private ChatCommandManager chatCommandManager;
 
 	@Inject
+	private TaskService taskService;
+	
+	@Inject
+	private SaveDataManager saveDataManager;
+
+	@Inject
+	private InterfaceManager interfaceManager;
+
+	@Inject
 	private ItemManager itemManager;
 
-	private SpriteDefinition[] spriteDefinitions;
-	private TieredTaskList tasks;
-
-	@Getter
-	private SaveData saveData;
-
-
-	private TaskDashboard taskDashboard;
-	private TaskList taskList;
-	private UICheckBox taskDashboardCheckbox;
-
-	private List<UIButton> tabs;
-	private UIButton taskListTab;
-	private UIButton taskDashboardTab;
-
 	private Map<Integer, Integer> chatSpriteMap = new HashMap<>();
-
-	private int activeTab = 0;
 
 	private File playerFile;
 
 	@Override
 	protected void startUp() throws Exception
 	{
-		this.spriteDefinitions = loadDefinitionResource(SpriteDefinition[].class, DEF_FILE_SPRITES, gson);
-		// Load the default task list so we definitely have something working
-		this.tasks = loadDefinitionResource(TieredTaskList.class, DEF_FILE_TASKS, gson);
-		if (config.loadRemoteTaskList()) {
-			loadRemoteTaskList();
-		}
-		this.spriteManager.addSpriteOverrides(spriteDefinitions);
 		mouseManager.registerMouseWheelListener(this);
+		interfaceManager.initialise();
 		this.taskOverlay.setResizable(true);
 		this.overlayManager.add(this.taskOverlay);
 		// TODO when task save data can be stored and access externally; populate this with other people's data
@@ -182,411 +102,56 @@ public class LogMasterPlugin extends Plugin implements MouseWheelListener
 //		chatCommandManager.registerCommandAsync(TASK_CHAT_COMMAND, this::getTaskCommandData);
 	}
 
-	private void populateChatSpriteMap() {
-		Set<Integer> itemIdsToLoad = new HashSet<>();
-		for (TaskTier tier : TaskTier.values()) {
-			itemIdsToLoad.addAll(this.tasks.getForTier(tier).stream().map(Task::getItemID).collect(Collectors.toList()));
-		}
-		List<Integer> itemIdsToLoadOrdered = new ArrayList<>(itemIdsToLoad);
-		final IndexedSprite[] modIcons = client.getModIcons();
-
-		final IndexedSprite[] newModIcons = Arrays.copyOf(modIcons, modIcons.length + itemIdsToLoadOrdered.size());
-		int modIconIdx = modIcons.length;
-
-		for (int i = 0; i < itemIdsToLoadOrdered.size(); i++)
-		{
-			final Integer itemId = itemIdsToLoadOrdered.get(i);
-			final ItemComposition itemComposition = itemManager.getItemComposition(itemId);
-			final BufferedImage image = ImageUtil.resizeImage(itemManager.getImage(itemComposition.getId()), 18, 16);
-			final IndexedSprite sprite = ImageUtil.getImageIndexedSprite(image, client);
-			final int spriteIndex = modIconIdx + i;
-
-			newModIcons[spriteIndex] = sprite;
-			chatSpriteMap.put(itemId, spriteIndex);
-		}
-
-		client.setModIcons(newModIcons);
-	}
-
-	private void getTaskCommandData(ChatMessage chatMessage, String message) {
-//		if (!config.taskChatCommand()) {
-//			return;
-//		}
-
-		Integer percentage = this.completionPercentages().get(this.getCurrentTier());
-
-		ChatMessageBuilder chatMessageBuilder =
-				new ChatMessageBuilder()
-						.append(ChatColorType.NORMAL)
-						.append("Progress: ")
-						.append(ChatColorType.HIGHLIGHT)
-						.append(percentage + "% " + this.getCurrentTier().displayName);
-
-		if (saveData.getActiveTaskPointer() != null) {
-			chatMessageBuilder
-					.append(ChatColorType.NORMAL)
-					.append(" Current task: ")
-					.img(chatSpriteMap.getOrDefault(saveData.getActiveTaskPointer().getTask().getItemID(), Integer.MIN_VALUE))
-					.append(ChatColorType.HIGHLIGHT)
-					.append(saveData.getActiveTaskPointer().getTask().getDescription());
-		} else {
-			chatMessageBuilder
-					.append(ChatColorType.NORMAL)
-					.append(" No current task.");
-		}
-
-		final String response = chatMessageBuilder.build();
-
-		final MessageNode messageNode = chatMessage.getMessageNode();
-		messageNode.setRuneLiteFormatMessage(response);
-		client.refreshChat();
-	}
-
-	protected Task currentTask() {
-		if (saveData == null || saveData.getActiveTaskPointer() == null) {
-			return null;
-		}
-		return saveData.getActiveTaskPointer().getTask();
-	}
-
-	protected void loadRemoteTaskList() {
-		log.info("Loading remote task list");
-		// Load the remote task list
-		try {
-			taskListClient.getTaskList(new Callback() {
-				@Override
-				public void onFailure(@NonNull Call call, @NonNull IOException e) {
-					log.error("Unable to load remote task list, will defer to the default task list", e);
-				}
-
-				@Override
-				public void onResponse(@NonNull Call call, @NonNull Response response) throws IOException {
-					JsonObject tasksJson = taskListClient.processResponse(response);
-					response.close();
-					if (tasksJson == null) {
-						log.error("Loaded null remote task list, will defer to the default task list");
-						return;
-					}
-
-					TieredTaskList tieredTaskList = gson.fromJson(tasksJson, TieredTaskList.class);
-					clientThread.invoke(() -> replaceTaskList(tieredTaskList));
-				}
-			});
-		} catch (IOException e) {
-			log.error("Unable to load remote task list, will defer to the default task list");
-		}
-	}
-
-	protected void replaceTaskList(TieredTaskList taskList) {
-		this.tasks = taskList;
-	}
-
 	@Override
-	protected void shutDown() throws Exception
-	{
+	protected void shutDown() throws Exception {
 		mouseManager.unregisterMouseWheelListener(this);
 	}
 
-	/**
-	 * Loads a definition resource from a JSON file
-	 *
-	 * @param classType the class into which the data contained in the JSON file will be read into
-	 * @param resource  the name of the resource (file name)
-	 * @param gson      a reference to the GSON object
-	 * @param <T>       the class type
-	 * @return the data read from the JSON definition file
-	 */
-	private <T> T loadDefinitionResource(Class<T> classType, String resource, Gson gson) {
-		// Load the resource as a stream and wrap it in a reader
-		InputStream resourceStream = classType.getResourceAsStream(resource);
-		assert resourceStream != null;
-		InputStreamReader definitionReader = new InputStreamReader(resourceStream);
-
-		// Load the objects from the JSON file
-		return gson.fromJson(definitionReader, classType);
-	}
-
-	/**
-	 * Sets up the playerFile variable, and makes the player file if needed.
-	 */
-	private void setupPlayerFile() {
-		saveData = new SaveData();
-
-		for (TaskTier loopTier : TaskTier.values()) {
-			if (!saveData.getProgress().containsKey(loopTier)) {
-				saveData.getProgress().put(loopTier, new HashSet<>());
-			}
-		}
-		File playerFolder = new File(RuneLite.RUNELITE_DIR, DATA_FOLDER_NAME);
-		if (!playerFolder.exists()) {
-			playerFolder.mkdirs();
-		}
-		playerFile = new File(playerFolder, client.getAccountHash() + ".txt");
-		if (!playerFile.exists()) {
-			try {
-				playerFile.createNewFile();
-			} catch (IOException e) {
-				e.printStackTrace();
-			}
-		} else {
-			loadPlayerData();
-		}
-	}
-
-	private void loadPlayerData() {
-		try {
-			String json = new Scanner(playerFile).useDelimiter("\\Z").next();
-			saveData = GSON.fromJson(json, new TypeToken<SaveData>() {}.getType());
-			for (TaskTier loopTier : TaskTier.values()) {
-				if (!saveData.getProgress().containsKey(loopTier)) {
-					saveData.getProgress().put(loopTier, new HashSet<>());
-				}
-			}
-			// Can get rid of this eventually
-			if (!this.saveData.getCompletedTasks().isEmpty()) {
-				this.saveData.getProgress().get(TaskTier.MASTER).addAll(this.saveData.getCompletedTasks().keySet());
-			}
-			if (saveData.currentTask != null) {
-				TaskPointer taskPointer = new TaskPointer();
-				taskPointer.setTask(saveData.currentTask);
-				taskPointer.setTaskTier(TaskTier.MASTER);
-				saveData.setActiveTaskPointer(taskPointer);
-				saveData.currentTask = null;
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-	private void savePlayerData() {
-		try {
-			PrintWriter w = new PrintWriter(playerFile);
-			String json = GSON.toJson(saveData);
-			w.println(json);
-			w.close();
-			log.debug("Saving player data");
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
 	@Subscribe
 	public void onConfigChanged(ConfigChanged event) {
 		if (!event.getGroup().equals("log-master")) {
 			return;
 		}
-		if (saveData == null) {
-			setupPlayerFile();
-		}
-		hideTabs();
-		updateTabs();
-		if (this.config.hideBelow() == TaskTier.MASTER && this.saveData.getSelectedTier() == TaskTier.MASTER && !this.taskDashboard.isVisible()) {
-			this.taskListTab.setSprites(TASKLIST_TAB_HOVER_SPRITE_ID);
-		}
-		if (this.taskDashboard != null && this.taskDashboardCheckbox.isEnabled()) {
-			showTabs();
-			if (this.saveData.getSelectedTier() != null && Arrays.asList(TaskTier.values()).indexOf(this.saveData.getSelectedTier()) < Arrays.asList(TaskTier.values()).indexOf(this.config.hideBelow())) {
-				activateTaskDashboard();
-			}
-			this.taskDashboard.updatePercentages();
-		}
+		interfaceManager.updateAfterConfigChange();
 	}
 
 	@Subscribe
-	public void onGameStateChanged(GameStateChanged gameStateChanged)
-	{
-		if (gameStateChanged.getGameState() == GameState.LOGGED_IN)
-		{
-			if(saveData == null) {
-				setupPlayerFile();
-			}
+	public void onGameStateChanged(GameStateChanged gameStateChanged) {
+		if (gameStateChanged.getGameState() == GameState.LOGGED_IN) {
+			saveDataManager.getSaveData();
 		} else if(gameStateChanged.getGameState().equals(GameState.LOGIN_SCREEN)) {
-			if(saveData != null) {
-				savePlayerData();
-			}
-
-			saveData = null;
+			saveDataManager.save();
 		}
 	}
 
 	@Subscribe
 	public void onWidgetLoaded(WidgetLoaded e) {
 		if(e.getGroupId() == WidgetInfo.COLLECTION_LOG.getGroupId()) {
-			if (saveData == null) {
-				setupPlayerFile();
-			}
-			Widget window = client.getWidget(40697857);
-
-			Widget dashboardTabWidget = window.createChild(-1, WidgetType.GRAPHIC);
-			taskDashboardTab = new UIButton(dashboardTabWidget);
-			taskDashboardTab.setSprites(DASHBOARD_TAB_SPRITE_ID, DASHBOARD_TAB_HOVER_SPRITE_ID);
-			taskDashboardTab.setSize(95, 21);
-			taskDashboardTab.setPosition(10, 36);
-			taskDashboardTab.addAction("View <col=ff9040>Dashboard</col>", this::activateTaskDashboard);
-			taskDashboardTab.setVisibility(false);
-
-			tabs = new ArrayList<>();
-
-			int currentTabX = 110;
-
-			for (int i = 0; i < 5; i++) {
-				Widget tabWiget = window.createChild(-1, WidgetType.GRAPHIC);
-				UIButton tab = new UIButton(tabWiget);
-				tab.setSize(66, 21);
-				tab.setPosition(currentTabX, 36);
-				tab.setVisibility(false);
-				currentTabX += 71;
-				tabs.add(tab);
-			}
-
-			Widget tabWiget = window.createChild(-1, WidgetType.GRAPHIC);
-			taskListTab = new UIButton(tabWiget);
-			taskListTab.setSize(95, 21);
-			taskListTab.setPosition(110, 36);
-			taskListTab.setSprites(TASKLIST_TAB_SPRITE_ID, TASKLIST_TAB_HOVER_SPRITE_ID);
-			taskListTab.setVisibility(false);
-			taskListTab.addAction("View <col=ff9040>Task List</col>", () -> {
-				taskDashboardTab.setSprites(DASHBOARD_TAB_SPRITE_ID, DASHBOARD_TAB_HOVER_SPRITE_ID);
-				if (this.saveData.getSelectedTier() != TaskTier.MASTER) {
-					this.taskList.goToTop();
-					this.saveData.setSelectedTier(TaskTier.MASTER);
-				}
-				updateTabs();
-				taskListTab.setSprites(TASKLIST_TAB_HOVER_SPRITE_ID);
-				this.taskDashboard.setVisibility(false);
-				this.taskList.refreshTasks(0);
-				this.taskList.setVisibility(true);
-			});
-
-			Widget dividerWidget = window.createChild(-1, WidgetType.GRAPHIC);
-			UIGraphic divider = new UIGraphic(dividerWidget);
-			divider.setSprite(DIVIDER_SPRITE_ID);
-			divider.setSize(480, 1);
-			divider.setPosition(10, 56);
-
-			createTaskDashboard(window);
-			createTaskList(window);
-			createTaskCheckbox();
-			updateTabs();
-
-			this.taskDashboardCheckbox.setEnabled(false);
-			this.taskDashboard.setVisibility(false);
-		}
-	}
-
-	private void updateTabs() {
-		int tabIndex = 0;
-		for (TaskTier tier : TaskTier.values()) {
-			if (tabIndex > 0 || tier == config.hideBelow()) {
-				if (tabs == null) {
-					return;
-				}
-				if (this.saveData.getSelectedTier() == tier && !this.taskDashboard.isVisible()) {
-					tabs.get(tabIndex).setSprites(tier.tabSpriteHoverId);
-				} else {
-					tabs.get(tabIndex).setSprites(tier.tabSpriteId, tier.tabSpriteHoverId);
-				}
-				int finalTabIndex = tabIndex;
-				tabs.get(tabIndex).clearActions();
-				tabs.get(tabIndex).setSize(66, 21);
-				tabs.get(tabIndex).addAction(String.format("View <col=ff9040>%s Task List</col>", tier.displayName), () -> {
-					taskDashboardTab.setSprites(DASHBOARD_TAB_SPRITE_ID, DASHBOARD_TAB_HOVER_SPRITE_ID);
-					if (this.saveData.getSelectedTier() != tier) {
-						this.taskList.goToTop();
-						this.saveData.setSelectedTier(tier);
-					}
-					updateTabs();
-					tabs.get(finalTabIndex).setSprites(tier.tabSpriteHoverId);
-					this.taskDashboard.setVisibility(false);
-					this.taskList.refreshTasks(0);
-					this.taskList.setVisibility(true);
-				});
-				tabIndex++;
-			}
+			interfaceManager.handleCollectionLogOpen();
 		}
 	}
 
 	@Subscribe
 	public void onWidgetClosed(WidgetClosed e) {
 		if(e.getGroupId() == WidgetInfo.COLLECTION_LOG.getGroupId()) {
-			this.taskDashboard.setVisibility(false);
-			this.taskList.setVisibility(false);
-			hideTabs();
-			this.taskDashboardCheckbox.setEnabled(false);
+			interfaceManager.handleCollectionLogClose();
 		}
 	}
 
 	@Subscribe
-	public void onGameTick(GameTick event)
-	{
-		if (this.taskList != null)
-			taskList.updateBounds();
+	public void onGameTick(GameTick event) {
+		interfaceManager.updateTaskListBounds();
 	}
 
 	@Override
-	public MouseWheelEvent mouseWheelMoved(MouseWheelEvent event)
-	{
-		if(this.taskList != null) {
-			taskList.handleWheel(event);
-		}
-
+	public MouseWheelEvent mouseWheelMoved(MouseWheelEvent event) {
+		interfaceManager.handleMouseWheel(event);
 		return event;
 	}
 
-	private void createTaskCheckbox() {
-		Widget window = client.getWidget(40697857);
-		// Create the graphic widget for the checkbox
-		Widget toggleWidget = window.createChild(-1, WidgetType.GRAPHIC);
-		Widget labelWidget = window.createChild(-1, WidgetType.TEXT);
-
-		// Wrap in checkbox, set size, position, etc.
-		taskDashboardCheckbox = new UICheckBox(toggleWidget, labelWidget);
-		taskDashboardCheckbox.setPosition(360, 10);
-		taskDashboardCheckbox.setName("Task Dashboard");
-		taskDashboardCheckbox.setEnabled(false);
-		taskDashboardCheckbox.setText("Task Dashboard");
-		labelWidget.setPos(375, 10);
-		taskDashboardCheckbox.setToggleListener(this::toggleTaskDashboard);
-	}
-
-	private void createTaskDashboard(Widget window) {
-		this.taskDashboard = new TaskDashboard(this, config, window);
-		this.taskDashboard.setVisibility(false);
-	}
-
-	private void createTaskList(Widget window) {
-		this.taskList = new TaskList(window, this.tasks, this, clientThread);
-		this.taskList.setVisibility(false);
-	}
-
-	private void toggleTaskDashboard(UIComponent src) {
-		if(this.taskDashboard == null) return;
-
-		if(saveData.getActiveTaskPointer() != null) {
-			this.taskDashboard.setTask(this.saveData.getActiveTaskPointer().getTask().getDescription(), this.saveData.getActiveTaskPointer().getTask().getItemID(), null);
-			this.taskDashboard.disableGenerateTask();
-		} else {
-			nullCurrentTask();
-		}
-
-		client.getWidget(COLLECTION_LOG_CONTENT_WIDGET_ID).setHidden(this.taskDashboardCheckbox.isEnabled());
-		client.getWidget(40697936).setHidden(this.taskDashboardCheckbox.isEnabled());
-
-		if (this.taskDashboardCheckbox.isEnabled()) {
-			activateTaskDashboard();
-		} else {
-			this.taskDashboard.setVisibility(false);
-			this.taskList.setVisibility(false);
-
-			hideTabs();
-		}
-
-		// *Boop*
-		this.client.playSoundEffect(SoundEffectID.UI_BOOP);
-	}
-
 	public void generateTask() {
-		if(this.saveData.currentTask != null || this.tasks == null) {
-			this.taskDashboard.disableGenerateTask();
+		if(this.saveDataManager.getSaveData().currentTask != null || taskService.getTaskList() == null) {
+			interfaceManager.disableGenerateTaskButton();
 			return;
 		}
 
@@ -606,43 +171,36 @@ public class LogMasterPlugin extends Plugin implements MouseWheelListener
 		TaskPointer newTaskPointer = new TaskPointer();
 		newTaskPointer.setTask(uniqueTasks.get(index));
 		newTaskPointer.setTaskTier(getCurrentTier());
-		this.saveData.setActiveTaskPointer(newTaskPointer);
-		this.taskDashboard.setTask(this.saveData.getActiveTaskPointer().getTask().getDescription(), this.saveData.getActiveTaskPointer().getTask().getItemID(), config.rollPastCompleted() ? this.tasks.getForTier(getCurrentTier()) : uniqueTasks);
-		log.debug("Task generated: "+this.saveData.getActiveTaskPointer().getTask().getDescription());
+		this.saveDataManager.getSaveData().setActiveTaskPointer(newTaskPointer);
+		interfaceManager.rollTask(this.saveDataManager.getSaveData().getActiveTaskPointer().getTask().getDescription(), this.saveDataManager.getSaveData().getActiveTaskPointer().getTask().getItemID(), config.rollPastCompleted() ? taskService.getForTier(getCurrentTier()) : uniqueTasks);
+		log.debug("Task generated: "+this.saveDataManager.getSaveData().getActiveTaskPointer().getTask().getDescription());
 
-		this.taskDashboard.disableGenerateTask(false);
-		taskList.refreshTasks(0);
-
-		this.taskDashboard.updatePercentages();
-		savePlayerData();
+		this.saveDataManager.save();
 	}
 
 	public void completeTask() {
-		completeTask(saveData.getActiveTaskPointer().getTask().getId(), saveData.getActiveTaskPointer().getTaskTier());
+		completeTask(saveDataManager.getSaveData().getActiveTaskPointer().getTask().getId(), saveDataManager.getSaveData().getActiveTaskPointer().getTaskTier());
 	}
 
 	public void completeTask(int taskID, TaskTier tier) {
 		this.client.playSoundEffect(SoundEffectID.UI_BOOP);
 
-		if (saveData.getProgress().get(tier).contains(taskID)) {
-			saveData.getProgress().get(tier).remove(taskID);
+		if (saveDataManager.getSaveData().getProgress().get(tier).contains(taskID)) {
+			saveDataManager.getSaveData().getProgress().get(tier).remove(taskID);
 		} else {
 			addCompletedTask(taskID, tier);
-			if (saveData.getActiveTaskPointer() != null && taskID == saveData.getActiveTaskPointer().getTask().getId()) {
+			if (saveDataManager.getSaveData().getActiveTaskPointer() != null && taskID == saveDataManager.getSaveData().getActiveTaskPointer().getTask().getId()) {
 				nullCurrentTask();
 			}
 		}
-		this.taskDashboard.updatePercentages();
+		interfaceManager.completeTask();
 
-		taskList.refreshTasks(0);
-
-		savePlayerData();
+		this.saveDataManager.save();
 	}
 
-	private void nullCurrentTask() {
-		this.saveData.setActiveTaskPointer(null);
-		this.taskDashboard.setTask("No task.", -1, null);
-		this.taskDashboard.enableGenerateTask();
+	public void nullCurrentTask() {
+		this.saveDataManager.getSaveData().setActiveTaskPointer(null);
+		interfaceManager.clearCurrentTask();
 	}
 
 	public static int getCenterX(Widget window, int width) {
@@ -654,20 +212,20 @@ public class LogMasterPlugin extends Plugin implements MouseWheelListener
 	}
 
 	public void addCompletedTask(int taskID, TaskTier tier) {
-		this.saveData.getProgress().get(tier).add(taskID);
+		this.saveDataManager.getSaveData().getProgress().get(tier).add(taskID);
 	}
 
 	public TaskTier getCurrentTier() {
-		if (this.saveData.getProgress().get(TaskTier.EASY).size() < this.tasks.getEasy().size() &&
+		if (this.saveDataManager.getSaveData().getProgress().get(TaskTier.EASY).size() < taskService.getTaskList().getEasy().size() &&
 				!Arrays.asList(TaskTier.MEDIUM, TaskTier.HARD, TaskTier.ELITE, TaskTier.MASTER).contains(config.hideBelow())) {
 			return TaskTier.EASY;
-		} else if (this.saveData.getProgress().get(TaskTier.MEDIUM).size() < this.tasks.getMedium().size() &&
+		} else if (this.saveDataManager.getSaveData().getProgress().get(TaskTier.MEDIUM).size() < taskService.getTaskList().getMedium().size() &&
 				!Arrays.asList(TaskTier.HARD, TaskTier.ELITE, TaskTier.MASTER).contains(config.hideBelow())) {
 			return TaskTier.MEDIUM;
-		} else if (this.saveData.getProgress().get(TaskTier.HARD).size() < this.tasks.getHard().size() &&
+		} else if (this.saveDataManager.getSaveData().getProgress().get(TaskTier.HARD).size() < taskService.getTaskList().getHard().size() &&
 				!Arrays.asList(TaskTier.ELITE, TaskTier.MASTER).contains(config.hideBelow())) {
 			return TaskTier.HARD;
-		} else if (this.saveData.getProgress().get(TaskTier.ELITE).size() < this.tasks.getElite().size() &&
+		} else if (this.saveDataManager.getSaveData().getProgress().get(TaskTier.ELITE).size() < taskService.getTaskList().getElite().size() &&
 				TaskTier.MASTER != config.hideBelow()) {
 			return TaskTier.ELITE;
 		} else {
@@ -676,70 +234,21 @@ public class LogMasterPlugin extends Plugin implements MouseWheelListener
 	}
 
 	public TaskTier getSelectedTier() {
-		return this.saveData.getSelectedTier();
+		return this.saveDataManager.getSaveData().getSelectedTier();
 	}
 
 	public List<Task> findAvailableTasks() {
-		return this.tasks.getForTier(getCurrentTier()).stream().filter(t -> !this.saveData.getProgress().get(getCurrentTier()).contains(t.getId())).collect(Collectors.toList());
-	}
-
-	private void hideTabs() {
-		if (this.taskDashboardTab != null) {
-			this.taskDashboardTab.setVisibility(false);
-		}
-		if (this.tabs != null) {
-			this.tabs.forEach(t -> t.setVisibility(false));
-		}
-		if (this.taskListTab != null) {
-			this.taskListTab.setVisibility(false);
-		}
-	}
-
-	private void showTabs() {
-		if (this.taskDashboardTab != null) {
-			this.taskDashboardTab.setVisibility(true);
-		}
-		if (this.config.hideBelow() == TaskTier.MASTER) {
-			this.taskListTab.setVisibility(true);
-		} else {
-			int tabIndex = 0;
-			for (TaskTier tier : TaskTier.values()) {
-				if (tabIndex > 0 || tier == config.hideBelow()) {
-					this.tabs.get(tabIndex).setVisibility(true);
-					tabIndex++;
-				}
-			}
-		}
-	}
-
-	private void activateTaskDashboard() {
-		this.taskDashboardTab.setSprites(DASHBOARD_TAB_HOVER_SPRITE_ID);
-		this.taskList.setVisibility(false);
-		this.taskDashboard.setVisibility(true);
-		this.taskListTab.setSprites(TASKLIST_TAB_SPRITE_ID, TASKLIST_TAB_HOVER_SPRITE_ID);
-		updateTabs();
-		showTabs();
+		return taskService.getTaskList().getForTier(getCurrentTier()).stream().filter(t -> !this.saveDataManager.getSaveData().getProgress().get(getCurrentTier()).contains(t.getId())).collect(Collectors.toList());
 	}
 
 	public void playFailSound() {
 		client.playSoundEffect(2277);
 	}
 
-	public Map<TaskTier, Integer> completionPercentages() {
-		Map<TaskTier, Integer> completionPercentages = new HashMap<>();
-		for (TaskTier tier : TaskTier.values()) {
-			completionPercentages.put(tier, (int) Math.round(((double) saveData.getProgress().get(tier).size() / (double) this.tasks.getForTier(tier).size()) * 100));
-		}
-		return completionPercentages;
-	}
 
 	@Provides
 	LogMasterConfig provideConfig(ConfigManager configManager)
 	{
 		return configManager.getConfig(LogMasterConfig.class);
-	}
-
-	public boolean isDashboardOpen() {
-		return this.taskDashboard != null && this.taskDashboard.isVisible();
 	}
 }
