+++ src/main/java/com/logmaster/ui/InterfaceConstants.java
@@ -1,0 +1,32 @@
+package com.logmaster.ui;
+
+public class InterfaceConstants {
+
+    public static final String DEF_FILE_SPRITES = "SpriteDef.json";
+
+    public static final int TASK_BACKGROUND_SPRITE_ID = -20006;
+    public static final int TASK_LIST_BACKGROUND_SPRITE_ID = -20012;
+    public static final int TASK_COMPLETE_BACKGROUND_SPRITE_ID = -20013;
+    public static final int TASK_CURRENT_BACKGROUND_SPRITE_ID = -20016;
+
+    public static final int COLLECTION_LOG_WINDOW_WIDTH = 500;
+    public static final int COLLECTION_LOG_WINDOW_HEIGHT = 314;
+    public static final int COLLECTION_LOG_CONTENT_WIDGET_ID = 40697858;
+
+    public static final int DASHBOARD_TAB_SPRITE_ID = -20007;
+    public static final int DASHBOARD_TAB_HOVER_SPRITE_ID = -20008;
+    public static final int TASKLIST_TAB_SPRITE_ID = -20009;
+    public static final int TASKLIST_TAB_HOVER_SPRITE_ID = -20010;
+    public static final int DIVIDER_SPRITE_ID = -20011;
+
+    public static final int TASKLIST_EASY_TAB_SPRITE_ID = -20017;
+    public static final int TASKLIST_EASY_TAB_HOVER_SPRITE_ID = -20018;
+    public static final int TASKLIST_MEDIUM_TAB_SPRITE_ID = -20019;
+    public static final int TASKLIST_MEDIUM_TAB_HOVER_SPRITE_ID = -20020;
+    public static final int TASKLIST_HARD_TAB_SPRITE_ID = -20021;
+    public static final int TASKLIST_HARD_TAB_HOVER_SPRITE_ID = -20022;
+    public static final int TASKLIST_ELITE_TAB_SPRITE_ID = -20023;
+    public static final int TASKLIST_ELITE_TAB_HOVER_SPRITE_ID = -20024;
+    public static final int TASKLIST_MASTER_TAB_SPRITE_ID = -20025;
+    public static final int TASKLIST_MASTER_TAB_HOVER_SPRITE_ID = -20026;
+}
