+++ src/main/java/com/logmaster/ui/generic/UIGraphic.java
@@ -1,4 +1,4 @@
-package com.logmaster.ui;
+package com.logmaster.ui.generic;
 
 import net.runelite.api.widgets.Widget;
