+++ src/main/resources/com/logmaster/ui/SpriteDef.json
