+++ src/main/resources/com/logmaster/ui/compTask-disabled.png
