+++ src/main/resources/com/logmaster/ui/compTask-hover.png
