+++ src/main/resources/com/logmaster/ui/compTask.png
