+++ src/main/resources/com/logmaster/ui/currentTask.png
