+++ src/main/resources/com/logmaster/ui/divider.png
