+++ src/main/resources/com/logmaster/ui/easy-tab-hover.png
