+++ src/main/resources/com/logmaster/ui/elite-tab.png
