+++ src/main/resources/com/logmaster/ui/genTask.png
