+++ src/main/resources/com/logmaster/ui/hard-tab-hover.png
