+++ src/main/resources/com/logmaster/ui/hard-tab.png
