+++ src/main/resources/com/logmaster/ui/master-tab-hover.png
