+++ src/main/resources/com/logmaster/ui/master-tab.png
