+++ src/main/resources/com/logmaster/ui/medium-tab-hover.png
