+++ src/main/resources/com/logmaster/ui/tab-dashboard-hover.png
