+++ src/main/resources/com/logmaster/ui/tab-dashboard.png
