+++ src/main/resources/com/logmaster/ui/task-long-complete.png
