+++ src/main/resources/com/logmaster/ui/task-long-current.png
