+++ src/main/resources/com/logmaster/ui/taskBg-complete.png
