+++ src/main/resources/com/logmaster/ui/up-arrow.png
